Team 8/requirements.txt
Pillow
pytesseract
openai 
pydantic
numpy
opencv-python
streamlit
python-dotenv
